requirements.txt
langchain
langchain-groq
python-dotenv
chromadb
langchain_community
beautifulsoup4
pymupdf
streamlit